refactor endpoint validation rules
